Very basic testing for complexity
Some small amount of a partial implementation has been
created. Currerntly works by exploring nested loops and
increasing the complexity count of the "node" or as I have
named it "analysis_node" which is a function.

In short, "analysis_node" is a function within the C program.
I check through how many nested loops there are.

To analyze a specific function, you run analyze_node on it, and then
the result of the function can be thought of as "n^{result from func}"
in terms of "big-o"