Complexity analysis solution
Complexity analysis approach was changed to a recursive
appproach (since we are dealing with a tree structure).

Approach is basically:
The weight (complexity) of some looping-node in our AST,
is equal to the heaviest child of that same looping-node.
So find the weights of all root looping-nodes of a function
and then identify the heaviest, and then return weight.

Colors was adjusted to make the output print a little cleaner
in terms of spacing. Old main.py was refactored to be more accurate
to what it is -- a proof of concept. It worked well for functions that
only had one root looping node, but didn't work for accurately finding
the complexity of the function if there was more root looping nodes.